update x and y to double

+++ dtos/elementdtos/ElementDto.cs
@@ -21,9 +21,9 @@ namespace webbuilder.api.dtos
         public Dictionary<string, object> Styles { get; init; } = [];
 
         [JsonPropertyName("x")]
-        public int X { get; init; }
+        public double X { get; init; }
         [JsonPropertyName("y")]
-        public int Y { get; init; }
+        public double Y { get; init; }
         [JsonPropertyName("src")]
         public string? Src { get; init; }
         [JsonPropertyName("href")]
